requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch
torchvision
opencv-python
mediapipe
numpy
matplotlib
pillow
